PR fixes
- Removed unnecessary classes.
- Removed moshi dependency from the delivery module.
- Use a custom request body to avoid reading the whole file every time we make a request
