+++ embrace-android-delivery/build.gradle.kts
@@ -12,7 +12,6 @@ apiValidation.validationDisabled = true
 
 dependencies {
     implementation(libs.okhttp)
-    implementation(libs.moshi)
     implementation(project(":embrace-android-payload"))
     implementation(project(":embrace-android-infra"))
 }
