+++ embrace-android-delivery/src/main/kotlin/io/embrace/android/embracesdk/internal/delivery/execution/RequestExecutionServiceImpl.kt
@@ -1,6 +1,6 @@
 package io.embrace.android.embracesdk.internal.delivery.execution
 
-import android.util.Log
+import io.embrace.android.embracesdk.internal.comms.api.ApiRequestV2
 import io.embrace.android.embracesdk.internal.comms.api.ApiResponse
 import io.embrace.android.embracesdk.internal.comms.api.Endpoint
 import io.embrace.android.embracesdk.internal.delivery.SupportedEnvelopeType
@@ -8,7 +8,10 @@ import okhttp3.Headers.Companion.toHeaders
 import okhttp3.OkHttpClient
 import okhttp3.Protocol
 import okhttp3.Request
-import okhttp3.RequestBody.Companion.toRequestBody
+import okhttp3.RequestBody
+import okio.BufferedSink
+import okio.buffer
+import okio.source
 import java.io.InputStream
 
 class RequestExecutionServiceImpl(
@@ -29,18 +32,23 @@ class RequestExecutionServiceImpl(
     ): ApiResponse {
         val apiRequest = envelopeType.endpoint.getApiRequestFromEndpoint()
 
-        val payload: ByteArray = payloadStream().use {
-            it.readBytes()
+        val requestBody = object : RequestBody() {
+            override fun contentType() = null
+
+            override fun writeTo(sink: BufferedSink) {
+                payloadStream().source().buffer().use { source ->
+                    sink.writeAll(source)
+                }
+            }
         }
 
         val request = Request.Builder()
             .url(apiRequest.url)
             .headers(apiRequest.getHeaders().toHeaders())
-            .post(payload.toRequestBody())
+            .post(requestBody)
             .build()
 
         okHttpClient.newCall(request).execute().use { response ->
-            Log.i("FRANEMBRACE", "response protocol: ${response.protocol}")
             if (response.isSuccessful) {
                 return ApiResponse.Success(null, null)
             }
@@ -59,4 +67,20 @@ class RequestExecutionServiceImpl(
         contentEncoding = "gzip",
         userAgent = "Embrace/a/$embraceVersionName"
     )
+
+    private fun ApiRequestV2.getHeaders(): Map<String, String> {
+        val headers = mutableMapOf(
+            "Accept" to accept,
+            "User-Agent" to userAgent,
+            "Content-Type" to contentType
+        )
+        contentEncoding?.let { headers["Content-Encoding"] = it }
+        acceptEncoding?.let { headers["Accept-Encoding"] = it }
+        appId?.let { headers["X-EM-AID"] = it }
+        deviceId?.let { headers["X-EM-DID"] = it }
+        eventId?.let { headers["X-EM-SID"] = it }
+        logId?.let { headers["X-EM-LID"] = it }
+        eTag?.let { headers["If-None-Match"] = it }
+        return headers
+    }
 }
